pyproject.toml
[tool.poetry]
name = "my-shared-infra"
version = "0.1.0"
description = "This repo will contain all my shared AWS infrastructure resources."
authors = ["thatsmidnight <10793854+thatsmidnight@users.noreply.github.com>"]
license = "Creative Commons Zero v1.0 Universal"
readme = "README.md"
packages = [{include = "my_shared_infra"}]

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.90.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"
coverage = "^7.2.7"
pytest-cov = "^4.1.0"
pylint = "^2.17.5"
poethepoet = "^0.21.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
py_version = 310
line_length = 79
wrap_length = 0
multi_line_output = 3
use_parentheses = true
combine_as_imports = true
include_trailing_comma = true
group_by_package = true
atomic = true
length_sort = true
import_heading_future = "Future"
import_heading_stdlib = "Standard Library"
import_heading_thirdparty = "Third Party"
import_heading_firstparty = "My Libraries"
import_heading_localfolder = "Local Modules"

[tool.black]
line-length = 79
target-version = ['py310']
